[Kernel] Expose QK RMSNorm and RoPE through shared diffusion ops
Move the existing implementation and registration without changing computation, retain the legacy public import, update H3 and tests, and document the current contract. Initial migration portion of RFC #7382 Phase 1; runtime and hardware qualification remain pending.
